Struct srml_timestamp::GenesisConfig

pub struct GenesisConfig<T: Trait> {
    pub minimum_period: T::Moment,
}

Fields
minimum_period: T::Moment

The minimum period between blocks. Beware that this is different to the expected period that the block production apparatus provides. Your chosen consensus system will generally work with this to determine a sensible block time. e.g. For Aura, it will be double this period on default settings.